move all signin logic to user provider
